# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Claudia Alzate Realtor® API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface AuthResult {
  username: string;
}

export interface CredentialsUpdate {
  username?: string;
  password?: string;
  currentPassword?: string;
}

export type ProfileVisualConfig = { [key: string]: unknown };

export interface Profile {
  name: string;
  subtitle: string;
  tagline: string;
  /** @nullable */
  logoUrl: string | null;
  /** @nullable */
  backgroundUrl: string | null;
  primaryColor: string;
  goldColor: string;
  fontTitle: string;
  fontBody: string;
  visualConfig?: ProfileVisualConfig;
}

export type ProfileUpdateVisualConfig = { [key: string]: unknown };

export interface ProfileUpdate {
  name?: string;
  subtitle?: string;
  tagline?: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  backgroundUrl?: string | null;
  primaryColor?: string;
  goldColor?: string;
  fontTitle?: string;
  fontBody?: string;
  visualConfig?: ProfileUpdateVisualConfig;
}

export interface Link {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  url: string;
  icon: string;
  active: boolean;
  order: number;
}

export interface LinkInput {
  title: string;
  /** @nullable */
  description?: string | null;
  url: string;
  icon: string;
  active?: boolean;
}

export interface LinkUpdate {
  title?: string;
  /** @nullable */
  description?: string | null;
  url?: string;
  icon?: string;
  active?: boolean;
  order?: number;
}

export interface LinksReorder {
  ids: number[];
}

export type TrackEventInputType = typeof TrackEventInputType[keyof typeof TrackEventInputType];


export const TrackEventInputType = {
  page_view: 'page_view',
  link_click: 'link_click',
} as const;

export interface TrackEventInput {
  type: TrackEventInputType;
  linkId?: number;
}

export interface AnalyticsLinkStat {
  title: string;
  clicks: number;
  icon: string;
}

export interface AnalyticsSummary {
  totalVisits: number;
  totalClicks: number;
  /** @nullable */
  mostClicked?: string | null;
  /** @nullable */
  lastActivity?: string | null;
  links: AnalyticsLinkStat[];
}

export type ResetAnalytics200 = {
  ok?: boolean;
};