# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Claudia Alzate Realtor® API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: profile
    description: Profile management
  - name: links
    description: Link management
  - name: analytics
    description: Analytics tracking and reporting
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Admin logout
      responses:
        "200":
          description: Logged out

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current session info
      responses:
        "200":
          description: Session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Not authenticated

  /auth/credentials:
    patch:
      operationId: updateCredentials
      tags: [auth]
      summary: Update username and/or password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialsUpdate"
      responses:
        "200":
          description: Updated
        "401":
          description: Not authenticated

  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get public profile data
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"

    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update profile (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Not authenticated

  /links:
    get:
      operationId: getLinks
      tags: [links]
      summary: Get all active links (public) or all links (admin)
      responses:
        "200":
          description: List of links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Link"

    post:
      operationId: createLink
      tags: [links]
      summary: Create a new link (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkInput"
      responses:
        "201":
          description: Created link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        "401":
          description: Not authenticated

  /links/{id}:
    patch:
      operationId: updateLink
      tags: [links]
      summary: Update a link (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkUpdate"
      responses:
        "200":
          description: Updated link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        "401":
          description: Not authenticated
        "404":
          description: Link not found

    delete:
      operationId: deleteLink
      tags: [links]
      summary: Delete a link (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Not authenticated
        "404":
          description: Link not found

  /analytics/track:
    post:
      operationId: trackEvent
      tags: [analytics]
      summary: Track a page view or link click (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackEventInput"
      responses:
        "200":
          description: Tracked

  /analytics:
    get:
      operationId: getAnalytics
      tags: [analytics]
      summary: Get analytics summary (admin only)
      responses:
        "200":
          description: Analytics summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsSummary"
        "401":
          description: Not authenticated

    delete:
      operationId: resetAnalytics
      tags: [analytics]
      summary: Delete all analytics data (admin only)
      responses:
        "200":
          description: Analytics reset
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "401":
          description: Not authenticated

  /links/reorder:
    patch:
      operationId: reorderLinks
      tags: [links]
      summary: Reorder links (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinksReorder"
      responses:
        "200":
          description: Reordered
        "401":
          description: Not authenticated

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    AuthResult:
      type: object
      required: [username]
      properties:
        username:
          type: string

    CredentialsUpdate:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        currentPassword:
          type: string

    Profile:
      type: object
      required: [name, subtitle, tagline, logoUrl, backgroundUrl, primaryColor, goldColor, fontTitle, fontBody]
      properties:
        name:
          type: string
        subtitle:
          type: string
        tagline:
          type: string
        logoUrl:
          type: ["string", "null"]
        backgroundUrl:
          type: ["string", "null"]
        primaryColor:
          type: string
        goldColor:
          type: string
        fontTitle:
          type: string
        fontBody:
          type: string
        visualConfig:
          type: object
          additionalProperties: true

    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        subtitle:
          type: string
        tagline:
          type: string
        logoUrl:
          type: ["string", "null"]
        backgroundUrl:
          type: ["string", "null"]
        primaryColor:
          type: string
        goldColor:
          type: string
        fontTitle:
          type: string
        fontBody:
          type: string
        visualConfig:
          type: object
          additionalProperties: true

    Link:
      type: object
      required: [id, title, url, icon, active, order]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        url:
          type: string
        icon:
          type: string
        active:
          type: boolean
        order:
          type: integer

    LinkInput:
      type: object
      required: [title, url, icon]
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        url:
          type: string
        icon:
          type: string
        active:
          type: boolean

    LinkUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        url:
          type: string
        icon:
          type: string
        active:
          type: boolean
        order:
          type: integer

    LinksReorder:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items:
            type: integer

    TrackEventInput:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [page_view, link_click]
        linkId:
          type: integer

    AnalyticsLinkStat:
      type: object
      required: [title, clicks, icon]
      properties:
        title:
          type: string
        clicks:
          type: integer
        icon:
          type: string

    AnalyticsSummary:
      type: object
      required: [totalVisits, totalClicks, links]
      properties:
        totalVisits:
          type: integer
        totalClicks:
          type: integer
        mostClicked:
          type: ["string", "null"]
        lastActivity:
          type: ["string", "null"]
        links:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsLinkStat"