# lib/api-zod/src/generated/types/analyticsLinkStat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Claudia Alzate Realtor® API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyticsLinkStat {
  title: string;
  clicks: number;
  icon: string;
}