# lib/api-zod/src/generated/types/analyticsSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Claudia Alzate Realtor® API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyticsLinkStat } from './analyticsLinkStat';

export interface AnalyticsSummary {
  totalVisits: number;
  totalClicks: number;
  /** @nullable */
  mostClicked?: string | null;
  /** @nullable */
  lastActivity?: string | null;
  links: AnalyticsLinkStat[];
}