# lib/api-zod/src/generated/types/credentialsUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Claudia Alzate Realtor® API
 * OpenAPI spec version: 0.1.0
 */

export interface CredentialsUpdate {
  username?: string;
  password?: string;
  currentPassword?: string;
}