# lib/api-zod/src/generated/types/link.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Claudia Alzate Realtor® API
 * OpenAPI spec version: 0.1.0
 */

export interface Link {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  url: string;
  icon: string;
  active: boolean;
  order: number;
}